support head_count_kv array in gguf
